split Utils into Currency genericUpgrade and upgrades
