+++ tsclient/src/OtherStuff/Data/GenericUpgrade.tsx
@@ -1,0 +1,10 @@
+import data from "../Util";
+
+export class upgrade extends data
+{
+    uButton=()=>{
+
+        this.addLvl();
+    }
+
+}
